Move RegexFileFilter to Core\Testing
