+++ Core/src/Testing/TestHelpers.php
@@ -18,13 +18,13 @@
 
 namespace Google\Cloud\Core\Testing;
 
+use Google\Cloud\Core\Testing\RegexFileFilter;
 use Google\Cloud\Core\Testing\Snippet\Container;
 use Google\Cloud\Core\Testing\Snippet\Coverage\Coverage;
 use Google\Cloud\Core\Testing\Snippet\Coverage\ExcludeFilter;
 use Google\Cloud\Core\Testing\Snippet\Coverage\Scanner;
 use Google\Cloud\Core\Testing\Snippet\Parser\Parser;
 use Google\Cloud\Core\Testing\System\SystemTestCase;
-use Google\Cloud\Dev\DocGenerator\RegexFileFilter;
 
 /**
  * Class TestHelpers is used to hold static functions required for testing
